When cannot get information from fb me, use the existing user info
Now when openning app without the internet access (previously logged in), the app shows a blank list. There issue was that we cannot get facebook user info as a result we couldn't get to the right user database. The fix is that if the user is previously logged in and we cannot get fb/me result, we will use the existing user info.